send dr to donwloadChan
