+++ zboxcore/sdk/downloadworker.go
@@ -31,7 +31,6 @@ import (
 	"go.dedis.ch/kyber/v3/group/edwards25519"
 	"golang.org/x/crypto/sha3"
 	"golang.org/x/sync/errgroup"
-	"golang.org/x/sync/semaphore"
 )
 
 const (
@@ -978,65 +977,4 @@ func (req *DownloadRequest) processDownloadRequest() {
 
 	blocksPerShard := (wantSize + int64(req.effectiveBlockSize) - 1) / int64(req.effectiveBlockSize)
 	req.blocksPerShard = blocksPerShard
-}
-
-func processReadMarker(drs []*DownloadRequest) {
-	blobberMap := make(map[uint64]int64)
-	mpLock := sync.Mutex{}
-	wg := sync.WaitGroup{}
-
-	for _, dr := range drs {
-		wg.Add(1)
-		go func(dr *DownloadRequest) {
-			defer wg.Done()
-			dr.processDownloadRequest()
-			var pos uint64
-			if !dr.skip {
-				for i := dr.downloadMask; !i.Equals64(0); i = i.And(zboxutil.NewUint128(1).Lsh(pos).Not()) {
-					pos = uint64(i.TrailingZeros())
-					mpLock.Lock()
-					blobberMap[pos] += dr.blocksPerShard
-					mpLock.Unlock()
-				}
-			}
-		}(dr)
-	}
-	wg.Wait()
-	successMask := zboxutil.NewUint128(0)
-	var redeemError error
-	for pos, totalBlocks := range blobberMap {
-		if totalBlocks == 0 {
-			continue
-		}
-		wg.Add(1)
-		go func(pos uint64, totalBlocks int64) {
-			blobber := drs[0].blobbers[pos]
-			err := drs[0].submitReadMarker(blobber, totalBlocks)
-			if err == nil {
-				successMask = successMask.Or(zboxutil.NewUint128(1).Lsh(pos))
-			} else {
-				redeemError = err
-			}
-			wg.Done()
-		}(pos, totalBlocks)
-	}
-	wg.Wait()
-	sem := semaphore.NewWeighted(downloadWorkerCount)
-	for _, dr := range drs {
-		if dr.skip {
-			continue
-		}
-		dr.downloadMask = successMask.And(dr.downloadMask)
-		if dr.consensusThresh > dr.downloadMask.CountOnes() {
-			dr.errorCB(redeemError, dr.remotefilepath)
-			continue
-		}
-		if err := sem.Acquire(dr.ctx, 1); err != nil {
-			continue
-		}
-		go func(dr *DownloadRequest) {
-			dr.processDownload(dr.ctx)
-			sem.Release(1)
-		}(dr)
-	}
 }
